+++ Assets/WaveformProvider/Resources.meta
@@ -1,0 +1,10 @@
+fileFormatVersion: 2
+guid: 502be8bb7a815f748bc73d93b358cba0
+folderAsset: yes
+timeCreated: 1516178922
+licenseType: Free
+DefaultImporter:
+  externalObjects: {}
+  userData: 
+  assetBundleName: 
+  assetBundleVariant:
