+++ Assets/WaveformProvider/Resources/Es.WaveformProvider.WaveInput.mat.meta
@@ -1,0 +1,10 @@
+fileFormatVersion: 2
+guid: 100fa6325bf41fd419941f544f83f240
+timeCreated: 1516172500
+licenseType: Free
+NativeFormatImporter:
+  externalObjects: {}
+  mainObjectFileID: 2100000
+  userData: 
+  assetBundleName: 
+  assetBundleVariant:
