+++ Assets/WaveformProvider/Resources/Es.WaveformProvider.WaveProvide.mat.meta
